Merging Generation Recipe (Mohamad's feature) with my previous features (adding ingredient to the meal plan)
Users can now add ingredients, which are saved in the Ingredient Overview. In the Recipe screen, if an ingredient is missing (e.g., tomato), adding it to the Ingredient Overview will automatically update the missing ingredients list, removing the previously missing item and adding it to the meal’s ingredients.